+++ lib/td_cache/tag.ex
@@ -1,0 +1,4 @@
+defmodule TdCache.Tag do
+  @moduledoc false
+  defstruct id: nil, type: nil, target_type: nil, expandable: nil, updated_at: nil
+end